[CHORE] : removed unused code
