+++ frontend/components/main/room_list/Rooms.tsx
@@ -52,7 +52,6 @@ export default function Rooms({
               setIsRight={setIsRight}
               setShowMembersList={setShowMembersList}
               isRight={isRight}
-              showMembersList={showMembersList}
               aRoom={aRoom}
             />
           );
